requirements.txt
ipykernel==5.5.6
numpy==1.26.4
pandas==2.2.3
matplotlib==3.7.5
scipy==1.13.1
seaborn==0.12.2
scikit-learn==1.2.2
transformers==4.47.0
torchmetrics==1.5.1
xgboost==2.0.3
